Reuse VerifyInvokeHTTPTests for GET requests
