+++ extensions/http-curl/tests/VerifyInvokeHTTPPostTest.cpp
@@ -15,106 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "VerifyInvokeHTTP.h"
 
-#undef NDEBUG
-#include "TestBase.h"
 #include "HTTPHandlers.h"
-#include "HTTPClient.h"
-#include "InvokeHTTP.h"
-#include "processors/LogAttribute.h"
-#include "core/state/ProcessorController.h"
-
-#include "CivetServer.h"
-#include "HTTPIntegrationBase.h"
 #include "utils/IntegrationTestUtils.h"
-
-class VerifyInvokeHTTP : public HTTPIntegrationBase {
- public:
-  VerifyInvokeHTTP()
-      : HTTPIntegrationBase(6000) {
-  }
-
-  void testSetup() override {
-    LogTestController::getInstance().setDebug<utils::HTTPClient>();
-    LogTestController::getInstance().setDebug<LogTestController>();
-    LogTestController::getInstance().setTrace<minifi::processors::InvokeHTTP>();
-    LogTestController::getInstance().setTrace<minifi::processors::LogAttribute>();
-  }
-
-  void setUrl(const std::string &url, ServerAwareHandler *handler) override {
-    if (path_) {
-      throw std::logic_error("Url is already set");
-    }
-    std::string port, scheme, path;
-    parse_http_components(url, port, scheme, path);
-    path_ = path;
-    HTTPIntegrationBase::setUrl(url, handler);
-  }
-
-  void setProperties(std::shared_ptr<core::Processor> proc) {
-    std::string url = scheme + "://localhost:" + getWebPort() + *path_;
-    proc->setProperty(minifi::processors::InvokeHTTP::URL.getName(), url);
-  }
-
-  void setProperty(const std::string& property, const std::string& value) {
-    const auto components = flowController_->getComponents("InvokeHTTP");
-    assert(!components.empty());
-
-    const auto stateController = components.at(0);
-    assert(stateController);
-    const auto processorController = std::dynamic_pointer_cast<minifi::state::ProcessorController>(stateController);
-    assert(processorController);
-    auto proc = processorController->getProcessor();
-    proc->setProperty(property, value);
-  }
-
-  virtual void setupFlow(const utils::optional<std::string>& flow_yml_path) {
-    testSetup();
-
-    std::shared_ptr<core::Repository> test_repo = std::make_shared<TestRepository>();
-    std::shared_ptr<core::Repository> test_flow_repo = std::make_shared<TestFlowRepository>();
-
-    if (flow_yml_path) {
-      configuration->set(minifi::Configure::nifi_flow_configuration_file, *flow_yml_path);
-    }
-    configuration->set("c2.agent.heartbeat.period", "200");
-    std::shared_ptr<core::ContentRepository> content_repo = std::make_shared<core::repository::VolatileContentRepository>();
-    content_repo->initialize(configuration);
-    std::shared_ptr<minifi::io::StreamFactory> stream_factory = minifi::io::StreamFactory::getInstance(configuration);
-    std::unique_ptr<core::FlowConfiguration> yaml_ptr = std::unique_ptr<core::YamlConfiguration>(
-        new core::YamlConfiguration(test_repo, test_repo, content_repo, stream_factory, configuration, flow_yml_path));
-
-    flowController_ = std::make_shared<minifi::FlowController>(test_repo, test_flow_repo, configuration, std::move(yaml_ptr), content_repo, DEFAULT_ROOT_GROUP_NAME, true);
-    flowController_->load();
-
-    std::string url = scheme + "://localhost:" + getWebPort() + *path_;
-    setProperty(minifi::processors::InvokeHTTP::URL.getName(), url);
-  }
-
-  void run(const utils::optional<std::string>& flow_yml_path = {}, const utils::optional<std::string>& = {}) override {
-    setupFlow(flow_yml_path);
-    startFlowController();
-
-    runAssertions();
-
-    shutdownBeforeFlowController();
-    stopFlowController();
-  }
-
-  void startFlowController() {
-    flowController_->start();
-  }
-
-  void stopFlowController() {
-    flowController_->unload();
-    flowController_->stopC2();
-
-    cleanup();
-  }
-
- private:
-  utils::optional<std::string> path_;
-};
 
 class VerifyInvokeHTTPOKResponse : public VerifyInvokeHTTP {
  public:
@@ -189,17 +93,6 @@ class VerifyRWTimeoutInvokeHTTP : public VerifyInvokeHTTP {
   }
 };
 
-void run(VerifyInvokeHTTP& harness,
-    const std::string& url,
-    const std::string& test_file_location,
-    const std::string& key_dir,
-    ServerAwareHandler * handler) {
-
-  harness.setKeyDir(key_dir);
-  harness.setUrl(url, handler);
-  harness.run(test_file_location);
-}
-
 int main(int argc, char ** argv) {
   const cmd_args args = parse_cmdline_args(argc, argv);
 
@@ -220,37 +113,37 @@ int main(int argc, char ** argv) {
   {
     InvokeHTTPResponseOKHandler handler;
     VerifyInvokeHTTPOKResponse harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   {
     InvokeHTTPRedirectHandler handler;
     VerifyInvokeHTTPOK200Response harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   {
     InvokeHTTPRedirectHandler handler;
     VerifyInvokeHTTPRedirectResponse harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   {
     InvokeHTTPResponse404Handler handler;
     VerifyNoRetryInvokeHTTP harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   {
     InvokeHTTPResponse501Handler handler;
     VerifyRetryInvokeHTTP harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   {
     TimeoutingHTTPHandler handler({std::chrono::seconds(2)});
     VerifyRWTimeoutInvokeHTTP harness;
-    run(harness, args.url, args.test_file, args.key_dir, &handler);
+    harness.run(args.url, args.test_file, args.key_dir, &handler);
   }
 
   return 0;
