+++ libminifi/test/resources/TestHTTPGetSecure.yml
@@ -20,7 +20,7 @@ Flow Controller:
     name: MiNiFi Flow
     id: 2438e3c8-015a-1000-79ca-83af40ec1990
 Processors:
-    - name: invoke
+    - name: InvokeHTTP
       id: 2438e3c8-015a-1000-79ca-83af40ec1991
       class: org.apache.nifi.processors.standard.InvokeHTTP
       max concurrent tasks: 1
@@ -55,7 +55,7 @@ Processors:
 Connections:
     - name: TransferFilesToRPG
       id: 2438e3c8-015a-1000-79ca-83af40ec1997
-      source name: invoke
+      source name: InvokeHTTP
       source id: 2438e3c8-015a-1000-79ca-83af40ec1991
       source relationship name: success
       destination name: LogAttribute
@@ -75,10 +75,10 @@ Connections:
       flowfile expiration: 60 sec
     - name: RetryInvokeHTTP
       id: c8a96f1d-48cd-4ee3-8a2f-081c3e3c7bcd
-      source name: invoke
+      source name: InvokeHTTP
       source id: 2438e3c8-015a-1000-79ca-83af40ec1991
       source relationship names: retry
-      destination name: invoke
+      destination name: InvokeHTTP
       destination id: 2438e3c8-015a-1000-79ca-83af40ec1991
       max work queue size: 0
       max work queue data size: 1 MB
